Support multimple registration certificates config
